refactor: Use a generic useFetch function to implement useGeoData
